myproj/requirements.txt
addict==2.2.1
cookiecutter==1.7.0
pyyaml==5.3
